requirements.txt
# requirements.txt
fastapi
uvicorn
requests
trafilatura
python-dotenv
pydantic

streamlit